Fix erroneous usage example in the docs
This never worked. The underlying reason is that wcsto... functions return zero both when an error is encountered, and when the input string was literally "0" (zero). This is unfortunate, because we can't differentiate, so when we see zero, we assume this is an error. Not nice but nothing tragic either, because 0 would mean "pass through", which is the default behavior anyway.